build fixes - update eslint package, move types around
