+++ .eslintrc.json
@@ -1,5 +1,5 @@
 {
-  "ignorePatterns": ["node_modules", "types", "coverage", "infra"],
+  "ignorePatterns": ["node_modules", "types", "coverage", "infra", "lambdas"],
   "env": {
     "es6": true
   },
@@ -47,6 +47,8 @@
     "@typescript-eslint/no-useless-constructor": "error",
     "@typescript-eslint/no-non-null-assertion": "warn",
     "@typescript-eslint/no-shadow": ["warn", { "builtinGlobals": true, "hoist": "all" }],
+    "no-unused-vars": "off",
+    "@typescript-eslint/no-unused-vars": ["error"],
     "@typescript-eslint/consistent-type-imports": [
       "error",
       {
