+++ components/portfolio/helpers/getPortfolioAccentColor.ts
@@ -1,4 +1,4 @@
-import type { PositionDetail } from 'features/portfolio/types'
+import type { PositionDetail } from 'lambdas/src/portfolio-positions/types'
 
 export function getPortfolioAccentColor(accent: PositionDetail['accent']): string {
   return accent === 'positive' ? 'success100' : 'error100'
