+++ components/portfolio/positions/PortfolioPositionsView.tsx
@@ -7,11 +7,12 @@ import { PortfolioPositionsSortingSelect } from 'components/portfolio/positions/
 import { PortfolioProductType, PortfolioSortingType } from 'components/portfolio/positions/types'
 import { Toggle } from 'components/Toggle'
 import { StatefulTooltip } from 'components/Tooltip'
-import type { PortfolioPositionsReply } from 'features/portfolio/types'
 import React, { useEffect, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Box, Flex, Grid, Text } from 'theme-ui'
 import { question_o } from 'theme/icons'
+
+import type { PortfolioPositionsReply } from 'lambdas/src/portfolio-positions/types'
 
 type PortfolioPositionsViewFiltersType = {
   showEmptyPositions: boolean
