+++ lambdas/src/portfolio-assets/types.ts
@@ -1,3 +1,37 @@
-import type { PortfolioAssetsReply } from 'features/portfolio/types'
+import type { NetworkNames } from 'blockchain/networks'
+export type DebankTokenReply = {
+  id: string
+  chain: string
+  name: string
+  symbol: string
+  display_symbol: string
+  optimized_symbol: string
+  decimals: number
+  logo_url: string
+  protocol_id: string
+  price: number
+  price_24h_change: number
+  is_verified: boolean
+  is_core: boolean
+  is_wallet: boolean
+  time_at: number
+  amount: number
+  raw_amount: number
+  raw_amount_hex_str: string
+}
+export type DebankTokensReply = DebankTokenReply[]
 
-export type PortfolioAssetsResponse = PortfolioAssetsReply
+export type PortfolioAssetsToken = {
+  name: string
+  network: NetworkNames
+  priceUSD: number
+  price24hChange: number
+  balance: number
+  balanceUSD: number
+}
+
+export type PortfolioAssetsReply = {
+  totalUSDAssets: number
+  totalUSDAssets24hChange: number
+  assets: PortfolioAssetsToken[]
+}
