+++ package.json
@@ -182,7 +182,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-simple-import-sort": "^5.0.2",
     "eslint-plugin-sonarjs": "^0.5.0",
-    "eslint-plugin-unused-imports": "^0.1.2",
+    "eslint-plugin-unused-imports": "3",
     "ethereum-types": "^3.3.3",
     "ethers": "5.7.0",
     "i18next-hmr": "^3.0.3",
