+++ tsconfig.test.json
@@ -6,5 +6,5 @@
     "lib": ["dom", "dom.iterable", "esnext"],
     "jsx": "react"
   },
-  "exclude": ["node_modules", "types", "coverage", "infra"]
+  "exclude": ["node_modules", "types", "coverage", "infra", "lambdas"]
 }
